chore(docs): Fix CNAME file being deleted
When I added the CNAME file to the `docs/` folder I forgot that we
delete the entire contents of the folder before building them. Since the
CNAME isn't autogenerated, that'd mean that it'd be deleted as soon as
somebody builds the docs.

This commit fixes that by copying the CNAME file from the root to the
`docs/` folder once the docs have been built.