Specify files to publish
This way not all folders and files get published. There is no need for the docs to be in the npm package.

The main reason why I submit this pr is because currently the `.babelrc` file is also published, and causing mayhem in my React Native project.

I'm not sure if the files I specified are sufficient, but it can be a start to cleaner publishes.