fix(mix): Fix first param issues with mix.
Introducing a breaking change with mix that now requires the weight in order to function. The
previous version that tried to make the first param optional with a fallback broke with Babel
compilation and currying, and also did not provide the original intended API.

BREAKING CHANGE: Mix no longer has a default weight, it must always be provided.